add base class to dynamic model of generator
