+++ src/main/groovy/com/powsybl/dynawo/dynamicmodels/GeneratorModelGroovyExtension.groovy
@@ -1,0 +1,43 @@
+/**
+ * Copyright (c) 2020, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.dynawo.dynamicmodels
+
+import java.util.function.Consumer
+
+import com.google.auto.service.AutoService
+import com.powsybl.dsl.DslException
+import com.powsybl.dynamicsimulation.DynamicModel
+import com.powsybl.dynamicsimulation.groovy.DynamicModelGroovyExtension
+
+/**
+ * @author Marcos de Miguel <demiguelm at aia.es>
+ */
+abstract class GeneratorModelGroovyExtension implements DynamicModelGroovyExtension {
+
+    static class GeneratorModelSpec {
+        String dynamicModelId
+        String staticId
+        String parameterSetId
+
+        void dynamicModelId(String dynamicModelId) {
+            this.dynamicModelId = dynamicModelId
+        }
+
+        void staticId(String staticId) {
+            this.staticId = staticId
+        }
+
+        void parameterSetId(String parameterSetId) {
+            this.parameterSetId = parameterSetId
+        }
+    }
+
+    String getName() {
+        return "Dynawo"
+    }
+    
+}
