+++ src/main/groovy/com/powsybl/dynawo/dynamicmodels/GeneratorSynchronousFourWindingsProportionalRegulationsGroovyExtension.groovy
@@ -19,47 +19,25 @@ import com.powsybl.dynamicsimulation.groovy.DynamicModelGroovyExtension
  * @author Marcos de Miguel <demiguelm at aia.es>
  */
 @AutoService(DynamicModelGroovyExtension.class)
-class GeneratorSynchronousFourWindingsProportionalRegulationsGroovyExtension implements DynamicModelGroovyExtension {
+class GeneratorSynchronousFourWindingsProportionalRegulationsGroovyExtension extends GeneratorModelGroovyExtension {
 
-    static class GeneratorSynchronousFourWindingsProportionalRegulationsSpec {
-        String dynamicModelId
-        String staticId
-        String parameterSetId
-
-        void dynamicModelId(String dynamicModelId) {
-            this.dynamicModelId = dynamicModelId
-        }
-
-        void staticId(String staticId) {
-            this.staticId = staticId
-        }
-
-        void parameterSetId(String parameterSetId) {
-            this.parameterSetId = parameterSetId
-        }
-    }
-
-    String getName() {
-        return "Dynawo"
-    }
-    
     void load(Binding binding, Consumer<DynamicModel> consumer) {
         binding.GeneratorSynchronousFourWindingsProportionalRegulations = { Closure<Void> closure ->
             def cloned = closure.clone()
-            GeneratorSynchronousFourWindingsProportionalRegulationsSpec generatorSynchronousFourWindingsProportionalRegulationsSpec = new GeneratorSynchronousFourWindingsProportionalRegulationsSpec()
+            GeneratorModelSpec generatorModelSpec = new GeneratorModelSpec()
 
-            cloned.delegate = generatorSynchronousFourWindingsProportionalRegulationsSpec
+            cloned.delegate = generatorModelSpec
             cloned()
 
-            if (!generatorSynchronousFourWindingsProportionalRegulationsSpec.staticId) {
+            if (!generatorModelSpec.staticId) {
                 throw new DslException("'staticId' field is not set")
             }
-            if (!generatorSynchronousFourWindingsProportionalRegulationsSpec.parameterSetId) {
+            if (!generatorModelSpec.parameterSetId) {
                 throw new DslException("'parameterSetId' field is not set")
             }
 
-            String dynamicModelId = generatorSynchronousFourWindingsProportionalRegulationsSpec.dynamicModelId ? generatorSynchronousFourWindingsProportionalRegulationsSpec.dynamicModelId : generatorSynchronousFourWindingsProportionalRegulationsSpec.staticId
-            consumer.accept(new GeneratorSynchronousFourWindingsProportionalRegulations(dynamicModelId, generatorSynchronousFourWindingsProportionalRegulationsSpec.staticId, generatorSynchronousFourWindingsProportionalRegulationsSpec.parameterSetId))
+            String dynamicModelId = generatorModelSpec.dynamicModelId ? generatorModelSpec.dynamicModelId : generatorModelSpec.staticId
+            consumer.accept(new GeneratorSynchronousFourWindingsProportionalRegulations(dynamicModelId, generatorModelSpec.staticId, generatorModelSpec.parameterSetId))
         }
     }
